# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is HireMeMaybe API service that provide data for HireMeMaybe web app",
        "title": "HireMeMaybe API service",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/jobpost": {
            "get": {
                "description": "Every query are not required, but they have specific use defined in their description",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobpost"
                ],
                "summary": "Get non-expired job posts based on query",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour access token\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search from job post title with substring matching and case insensitive",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Job type field with substring matching and case insensitive",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search if tags field contain tag param, no substring matching and case insensitive",
                        "name": "tag",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Salary field, must exactly match to get result",
                        "name": "salary",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Exp_lvl field, must exactly match to get result",
                        "name": "exp",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search from company name with substring matching and case insensitive",
                        "name": "company",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search from industry of company with substring matching and case insensitive",
                        "name": "industry",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search from location with substring matching and case insensitive",
                        "name": "location",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Sorting by post time in descending if true, otherwise ascendind",
                        "name": "desc",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return non-expired job post(s)",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.JobPost"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header, or invalid job post struct",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User is banned",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Only verified company have access to this endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobpost"
                ],
                "summary": "Create job post based on given json structure",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour access token\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Input jobpost information",
                        "name": "Jobpost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EditableJobPostInfo"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully create job post",
                        "schema": {
                            "$ref": "#/definitions/model.JobPost"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header, or invalid job post struct",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Not logged in as verified company, User is banned or suspended",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/jobpost/{id}": {
            "put": {
                "description": "Only company that own the post or admin have access to this endpoint",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobpost"
                ],
                "summary": "Edit job post based on given json structure",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour access token\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of desired job post",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Input jobpost information",
                        "name": "Jobpost",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.EditableJobPostInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully update job post",
                        "schema": {
                            "$ref": "#/definitions/model.JobPost"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header, or invalid job post struct",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Do not have permission to edit, User is banned",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Only company that own the post or admin have access to this endpoint",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Jobpost"
                ],
                "summary": "Delete given job post ID",
                "parameters": [
                    {
                        "type": "string",
                        "default": "Bearer \u003cyour access token\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of desired job post",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully delete job post",
                        "schema": {
                            "$ref": "#/definitions/utilities.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid authorization header, or invalid job post struct",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Do not have permission to delete this post, User is banned",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Database error",
                        "schema": {
                            "$ref": "#/definitions/utilities.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Application": {
            "type": "object",
            "required": [
                "post_id",
                "resume_id"
            ],
            "properties": {
                "answer": {
                    "$ref": "#/definitions/model.ApplicationAnswer"
                },
                "answer_id": {
                    "type": "integer"
                },
                "applied_at": {
                    "type": "string"
                },
                "cpsk_id": {
                    "description": "CPSKID references CPSKUser.UserID (uuid)",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "post_id": {
                    "description": "PostID references JobPost.ID",
                    "type": "integer"
                },
                "resume_id": {
                    "type": "integer"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "model.ApplicationAnswer": {
            "type": "object",
            "properties": {
                "expected_salary": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "programming_languages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "right_to_work": {
                    "type": "string"
                },
                "year_of_experience": {
                    "type": "integer"
                }
            }
        },
...
        "model.EditableJobPostInfo": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "exp_lvl": {
                    "type": "string"
                },
                "expiring": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "req": {
                    "type": "string"
                },
                "salary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "model.JobPost": {
            "type": "object",
            "properties": {
                "applications": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Application"
                    }
                },
                "company_id": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "exp_lvl": {
                    "type": "string"
                },
                "expiring": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "post_time": {
                    "type": "string"
                },
                "req": {
                    "type": "string"
                },
                "salary": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "utilities.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "utilities.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    }
}